changed current char default type to empty string
the _make_text() in lexer.py used to raise a TypeError because you can't compare None to string so i made it that the advance method doesn't set self.current_char to None instead empty string. another thing that i forgot to do was call the advance() method in the init method of lexer so added that as well.